requirements.txt
discord.py
plexapi
nest-asyncio